requirements.txt
appdirs==1.4.4
asgiref==3.5.2
azure-core==1.26.3
azure-identity==1.12.0
azure-storage-blob==12.15.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==39.0.2
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==0.5.0
Django==4.1.7
django-allauth==0.49.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-recaptcha==3.0.0
django-rest-auth==0.9.5
django-updown-ratings==1.0.1
django-widget-tweaks==1.4.3
djangorestframework==3.13.1
dpkt==1.9.8
filelock==3.9.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
lxml==4.9.3
Markdown==3.3.4
martor==1.6.12
mccabe==0.6.1
msal==1.21.0
msal-extensions==1.0.0
mypy-extensions==0.4.3
oauthlib==3.2.2
packaging==23.2
pathspec==0.9.0
Pillow==9.1.1
pipenv==2023.2.18
platformdirs==2.5.2
portalocker==2.7.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pygame==2.5.0
PyJWT==2.6.0
pyshark==0.6
python-decouple==3.1
python-memcached==1.59
python3-openid==3.2.0
pytz==2018.5
requests==2.27.1
requests-oauthlib==1.3.1
scapy==2.5.0
sentry-sdk==1.5.6
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.4.2
StackAPI==0.2.0
termcolor==2.3.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.14
virtualenv==20.19.0
virtualenv-clone==0.5.7
whitenoise==6.1.0
